tıklayınca forma lat lng girilmesi

+++ Models/PollutionData.cs
@@ -3,8 +3,8 @@
     public class PollutionData
     {
         public int Id { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
         public string City { get; set; }
         public string Region { get; set; }
         public string MetalType { get; set; }
